+++ src/math/Vector.h
@@ -5,23 +5,23 @@
 class Vector
 {
 public:
-	double x, y, z;
+	Float x, y, z;
 
 	Vector() : x{0}, y{0}, z{0} {}
-	Vector(double x, double y, double z) : x{x}, y{y}, z{z} {}
+	Vector(Float x, Float y, Float z) : x{x}, y{y}, z{z} {}
 
-	inline			double& operator[](unsigned);
-	inline const	double& operator[](unsigned) const;
+	inline			Float& operator[](unsigned);
+	inline const	Float& operator[](unsigned) const;
 	
 			Vector	operator- ()		{return {-x, -y, -z};}
 	const	Vector	operator- () const	{return {-x, -y, -z};}
 
-			Vector	operator+ (double a) {return {x+a, y+a, z+a};}
-			Vector	operator- (double a) {return (*this)+(-a);}
-			Vector	operator* (double a) {return {x*a, y*a, z*a};}
-			void	operator+=(double a) {x += a; y += a; z += a;}
-			void	operator-=(double a) {*this += -a;}
-			void	operator*=(double a) {x *= a; y *= a; z *= a;}
+			Vector	operator+ (Float a) {return {x+a, y+a, z+a};}
+			Vector	operator- (Float a) {return (*this)+(-a);}
+			Vector	operator* (Float a) {return {x*a, y*a, z*a};}
+			void	operator+=(Float a) {x += a; y += a; z += a;}
+			void	operator-=(Float a) {*this += -a;}
+			void	operator*=(Float a) {x *= a; y *= a; z *= a;}
 
 			Vector	operator+ (const Vector& v) {return {x+v.x, y+v.y, z+v.z};}
 			Vector	operator- (const Vector& v) {return (*this)+(-v);}
@@ -29,28 +29,28 @@ public:
 			void	operator-=(const Vector& v) {*this += -v;}
 
 
-	inline static double dot(const Vector&, const Vector&);
+	inline static Float dot(const Vector&, const Vector&);
 	inline static Vector cross(const Vector&, const Vector&);
 	inline static Vector normalize(const Vector&);
 
 	// Builds a reference system from a single vector
 	static std::pair<Vector, Vector> referenceSystem(const Vector&);
 };
-inline Vector operator*(double a, Vector v) {return v*a;}
+inline Vector operator*(Float a, Vector v) {return v*a;}
 
 
 
-double& Vector::operator[](unsigned idx)
+Float& Vector::operator[](unsigned idx)
 {
 	switch(idx) {case 0: return x; case 1: return y; case 2: return z;}
 }
 
-const double& Vector::operator[](unsigned idx) const
+const Float& Vector::operator[](unsigned idx) const
 {
 	switch(idx) {case 0: return x; case 1: return y; case 2: return z;}
 }
 
-double Vector::dot(const Vector& v1, const Vector& v2)
+Float Vector::dot(const Vector& v1, const Vector& v2)
 {
 	return v1.x*v2.x + v1.y*v2.y + v1.z*v2.z;
 }
@@ -62,7 +62,7 @@ Vector Vector::cross(const Vector& v1, const Vector& v2)
 
 Vector Vector::normalize(const Vector& v)
 {
-	double invLength = 1/std::sqrt(v.x*v.x + v.y*v.y + v.z*v.z);
+	Float invLength = 1/std::sqrt(v.x*v.x + v.y*v.y + v.z*v.z);
 	return invLength*v;
 }
